Generalized structure for all QEDC benchmarks.
