+++ metriq_gym/benchmarks/__init__.py
@@ -7,7 +7,7 @@ from metriq_gym.benchmarks.quantum_volume import QuantumVolume, QuantumVolumeDat
 from metriq_gym.benchmarks.bseq import BSEQ, BSEQData
 from metriq_gym.benchmarks.mirror_circuits import MirrorCircuits, MirrorCircuitsData
 from metriq_gym.benchmarks.wormhole import Wormhole, WormholeData
-from metriq_gym.benchmarks.bernstein_vazirani import BernsteinVazirani, BernsteinVaziraniData
+from metriq_gym.benchmarks.qedc_benchmarks import QEDCBenchmarks, QEDCData
 
 
 class JobType(StrEnum):
@@ -27,7 +27,7 @@ BENCHMARK_HANDLERS: dict[JobType, type[Benchmark]] = {
     JobType.QUANTUM_VOLUME: QuantumVolume,
     JobType.MIRROR_CIRCUITS: MirrorCircuits,
     JobType.WORMHOLE: Wormhole,
-    JobType.BERNSTEIN_VAZIRANI: BernsteinVazirani,
+    JobType.BERNSTEIN_VAZIRANI: QEDCBenchmarks,
 }
 
 BENCHMARK_DATA_CLASSES: dict[JobType, type[BenchmarkData]] = {
@@ -37,7 +37,7 @@ BENCHMARK_DATA_CLASSES: dict[JobType, type[BenchmarkData]] = {
     JobType.QUANTUM_VOLUME: QuantumVolumeData,
     JobType.MIRROR_CIRCUITS: MirrorCircuitsData,
     JobType.WORMHOLE: WormholeData,
-    JobType.BERNSTEIN_VAZIRANI: BernsteinVaziraniData,
+    JobType.BERNSTEIN_VAZIRANI: QEDCData,
 }
 
 SCHEMA_MAPPING = {
